+++ lib/components/dialogs.dart
@@ -1,9 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:past_paper_master/components/button.dart';
 import 'package:past_paper_master/core/colors.dart';
+import 'package:past_paper_master/core/provider.dart';
 import 'package:past_paper_master/core/textstyle.dart';
 import 'package:markdown_widget/widget/all.dart';
 import 'package:past_paper_master/core/global.dart';
+import 'package:past_paper_master/pages/about.dart';
+import 'package:provider/provider.dart';
 
 class MAlertDialogNoDownloadPath extends StatelessWidget {
   const MAlertDialogNoDownloadPath({
@@ -50,6 +53,54 @@ Future<dynamic> showReleaseNotesDialog(BuildContext context) {
           title: 'Dismiss',
           onPressed: () {
             Navigator.of(context).pop();
+          },
+        ),
+      ],
+    ),
+  );
+}
+
+Future<dynamic> showUpdateDialog(BuildContext context, String versionTag) {
+  return showDialog(
+    context: context,
+    builder: (context) => AlertDialog(
+      backgroundColor: MColors.white,
+      titleTextStyle: MTextStyles.lgMdGrey900,
+      shape: RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(12),
+      ),
+      title: const Text("Update available"),
+      content: Column(
+        mainAxisSize: MainAxisSize.min,
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          const Text("A new version of Past Paper Master is available!"),
+          Divider(
+            color: MColors.grey.shade300,
+            height: 30,
+          ),
+          Text(
+            'Current version: $kVersionTag',
+            style: MTextStyles.smRgGrey500,
+          ),
+          Text(
+            'Latest version: $versionTag',
+            style: MTextStyles.smMdAccent700,
+          ),
+        ],
+      ),
+      actions: [
+        MButton(
+          title: 'Dismiss',
+          onPressed: () {
+            Navigator.of(context).pop();
+          },
+        ),
+        MButton(
+          title: 'Download on GitHub',
+          primary: true,
+          onPressed: () {
+            safeLaunchUrl(context.read<SettingsCN>().latestReleaseUrl);
           },
         ),
       ],