+++ lib/core/global.dart
@@ -10,15 +10,21 @@ late BuildContext globalContext;
 const String kAppStage = 'Beta Version';
 
 const String kReleaseNotes = '''
+- Adds update checker
+  - Automatically checks for update
+  - Release notes will be shown after updates
+- Adds Clear Selection and Select All button to **Browse** page
+- Fixes download item cancelling
 - Fixes minor layout issues
-- Fixes subject search field
-- Reverts to Material 2 design
+  - Reverts to Material 2 design
+  - Fixes subject search field
+- Enhances app performance
 
 See [full changelog](https://github.com/SCIEDEV/PastPaperMaster/releases) and [commit history](https://github.com/SCIEDEV/PastPaperMaster/commits/main/) for changes in previous versions.
 ''';
 
 const String kAppStageShort = 'β ';
-const String kVersionTag = 'v0.1.7-beta';
+const String kVersionTag = 'v0.1.7-beta+65';
 const int kMajorVersion = 0;
 const int kMinorVersion = 1;
 const int kPatchVersion = 7;